Update README.md
This is a code made to simulate a trivia game within the console. The console randomly chooses a question and then the user writes the answer he thinks is correct. After that the console will tell you if you got the answer correct or if you got it wrong. And at last it will give you a score out of 15, if you have a score of 11 or less you will fail.